+++ packages/features/users/components/UserTable/EditSheet/store.ts
@@ -1,0 +1,11 @@
+import { create } from "zustand";
+
+interface EditModeState {
+  editMode: boolean;
+  setEditMode: (editMode: boolean) => void;
+}
+
+export const useEditMode = create<EditModeState>((set) => ({
+  editMode: false,
+  setEditMode: (editMode) => set({ editMode }),
+}));
